refactor: migrate DataInputStream and DataOutputStream to stream
